+++ packages/test-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix require for `yarn render`
 
 ## 1.0.3 - 2021/5/24